refactor: de-vendor chainaim namespace into standard layout
