--- packages/nest-core/nest_core/chainaim/__init__.py
@@ -1,24 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""ChainAIM outcome-verified-settlement validators package.
-
-Example::
-
-    from nest_core.chainaim.outcome_verified_settlement_validator import (
-        validate_outcome_verified_settlement_no_drain_after_close,
-        validate_outcome_verified_settlement_no_overbill,
-    )
-"""
-
-from __future__ import annotations
-
-from nest_core.chainaim.outcome_verified_settlement_validator import (
-    validate_outcome_verified_settlement_no_drain_after_close,
-    validate_outcome_verified_settlement_no_overbill,
-    validate_outcome_verified_settlement_no_overbill_on_failed_verification,
-)
-
-__all__ = [
-    "validate_outcome_verified_settlement_no_drain_after_close",
-    "validate_outcome_verified_settlement_no_overbill",
-    "validate_outcome_verified_settlement_no_overbill_on_failed_verification",
-]
