+++ packages/nest-core/nest_core/plugins.py
@@ -37,7 +37,7 @@ _BUILTINS: dict[tuple[str, str], str] = {
     # Problem 03: spec-suggested key ("payments", "streaming") is held by merged PR #21;
     # distinct key per the charter's anti-duplicate rule -- see the plugin README.
     ("payments", "outcome_verified_settlement"): (
-        f"{_REF}.payments.chainaim.outcome_verified_settlement:OutcomeVerifiedSettlement"
+        f"{_REF}.payments.outcome_verified_settlement:OutcomeVerifiedSettlement"
     ),
     ("coordination", "contract_net"): f"{_REF}.coordination.contract_net:ContractNet",
     ("coordination", "hotstuff"): f"{_REF}.coordination.hotstuff:HotStuff",
