+++ packages/nest-core/nest_core/scenarios.py
@@ -146,7 +146,7 @@ def _try_load_builtin(name: str) -> None:
 
         register_scenario("rogue_trusted_agent", rogue_trusted_agent_factory)
     elif name == "outcome_verified_settlement":
-        from nest_core.scenarios_builtin.chainaim.outcome_verified_settlement import (
+        from nest_core.scenarios_builtin.outcome_verified_settlement import (
             outcome_verified_settlement_factory,
         )
