--- packages/nest-core/nest_core/scenarios_builtin/chainaim/__init__.py
@@ -1,23 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""ChainAIM outcome-verified settlement scenario package.
-
-Example::
-
-    from nest_core.scenarios_builtin.chainaim.outcome_verified_settlement import (
-        outcome_verified_settlement_factory,
-    )
-"""
-
-from __future__ import annotations
-
-from nest_core.scenarios_builtin.chainaim.outcome_verified_settlement import (
-    OutcomeVerifiedSettlementBuyerAgent,
-    OutcomeVerifiedSettlementSellerAgent,
-    outcome_verified_settlement_factory,
-)
-
-__all__ = [
-    "OutcomeVerifiedSettlementBuyerAgent",
-    "OutcomeVerifiedSettlementSellerAgent",
-    "outcome_verified_settlement_factory",
-]
