+++ packages/nest-core/tests/test_outcome_verified_settlement_b1_gates.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 import hashlib
 
 import pytest
-from nest_core.scenarios_builtin.chainaim.gates import (
+from nest_core.scenarios_builtin.gates import (
     AckReceivedGate,
     ChecksumGate,
     Gate,
