+++ packages/nest-core/tests/test_outcome_verified_settlement_b3_validator.py
@@ -12,13 +12,14 @@ from __future__ import annotations
 
 from typing import Any
 
-from nest_core.chainaim.outcome_verified_settlement_validator import (
+from nest_core.validators import (
+    VALIDATORS,
+    validate_events,
     validate_outcome_verified_settlement_no_drain_after_close,
     validate_outcome_verified_settlement_no_overbill,
     validate_outcome_verified_settlement_no_overbill_on_failed_verification,
     validate_outcome_verified_settlement_verdicts_match_committed_criterion,
 )
-from nest_core.validators import VALIDATORS, validate_events
 
 type Event = dict[str, Any]
