--- packages/nest-plugins-reference/nest_plugins_reference/payments/chainaim/__init__.py
@@ -1,16 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""ChainAIM outcome-verified settlement plugin package.
-
-Example::
-
-    from nest_plugins_reference.payments.chainaim import StreamHandle, OutcomeVerifiedSettlement
-"""
-
-from __future__ import annotations
-
-from nest_plugins_reference.payments.chainaim.outcome_verified_settlement import (
-    OutcomeVerifiedSettlement,
-    StreamHandle,
-)
-
-__all__ = ["OutcomeVerifiedSettlement", "StreamHandle"]
