+++ packages/nest-plugins-reference/pyproject.toml
@@ -54,7 +54,7 @@ parc = "nest_plugins_reference.trust.parc:ParcTrust"
 aae_permit_gate = "nest_plugins_reference.trust.aae_permit_gate:AAEPermitGate"
 
 [project.entry-points."nest.plugins.payments"]
-outcome_verified_settlement = "nest_plugins_reference.payments.chainaim.outcome_verified_settlement:OutcomeVerifiedSettlement"
+outcome_verified_settlement = "nest_plugins_reference.payments.outcome_verified_settlement:OutcomeVerifiedSettlement"
 
 [build-system]
 requires = ["hatchling"]
